Add tests for tempalte.
